Fix java.lang.IllegalArgumentException only on Android 10 if multi dot
